refactor by CR
